refactor(streams): combine NStreams and NStreamsUtils
