--- src/main/java/com/nitorcreations/streams/NStreamsUtils.java
@@ -1,12 +1,0 @@
-package com.nitorcreations.streams;
-
-import java.util.stream.Stream;
-
-public class NStreamsUtils {
-
-    @SafeVarargs
-    public static <T> Stream<T> concat(Stream<T> ...streams) {
-      return Stream.of(streams).reduce(Stream::concat).get();
-    }
-
-}
